requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
neo4j
python-dotenv
